Implement full json api specifications

+++ app/controllers/v1/comments_controller.rb
@@ -3,7 +3,7 @@ module V1
     include Trackable
     include Notifiable
 
-    before_action :authenticate_with_token, except: %i[index]
+    before_action :authenticate_with_token, only: %i[create destroy]
 
     after_action  -> { create_activity(subject: find_resource,
                                        user: current_user,
@@ -17,31 +17,22 @@ module V1
     def index
       resource = find_resource
       comments = resource.comments
-      resources = {
-          user: {
-              fields: []
-          }
-      }
-      render json: CommentSerializer.new(comments, SerializationOption.run(resources)).serialized_json,
-             status: :ok
+      render json: JSONAPI::ResourceSerializer.new(CommentResource).
+          serialize_to_hash(CommentResource.new(comments, context)), status: :ok
     end
     
     def create
       resource = find_resource
-      comment = resource.comments.new(body: params[:body])
+      comment = resource.comments.new(body: params[:data][:attributes][:body])
       comment.user = current_user
       comment.save!
-      resources = {
-		      user: {
-				      fields: []
-		      }
-      }
-      render json: CommentSerializer.new(comment, SerializationOption.run(resources)).serialized_json, status: :created
+      render json: JSONAPI::ResourceSerializer.new(CommentResource).
+          serialize_to_hash(CommentResource.new(comment, context)), status: :created
     end
 
     def destroy
-      resource = find_resource
-      comment = resource.comments.find_by!(user: current_user, id: params[:id])
+      comment = Comment.find_by!(id: params[:id])
+      authorize comment
       comment.destroy
       head(:no_content)
     end
@@ -49,13 +40,8 @@ module V1
     private
 
     def find_resource
-      if params[:story_id].present?
-        Story.find_by!(id: params[:story_id])
-      elsif params[:creation_id].present?
-        Creation.find_by!(id: params[:creation_id])
-      elsif params[:project_id].present?
-        Project.find_by!(id: params[:project_id])
-      end
+      raise ActiveRecord::RecordNotFound unless params[:creation_id]
+      Creation.find_by!(id: params[:creation_id])
     end
 
     def activity_type
